# .choreo/openapi.yaml
openapi: 3.0.0
info:
  title: AD SOUTHERN SMART POS API
  version: 1.0.0
  description: >
    Full route map for the AD SOUTHERN SMART POS backend. This file exists so
    Choreo's gateway has an explicit definition of every path the Express app
    exposes. Without this, routes added after the component's first deploy
    (e.g. verify-pin) can fail to be wired into the gateway's route table,
    causing requests to hang until timeout instead of reaching the backend
    pod (visible in Gateway logs as ServiceHost="-").
servers:
  - url: /
paths:
...
  /api/admin/dashboard:
    get:
      operationId: adminDashboard
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Admin dashboard summary
...
  /api/admin/cheques:
    get:
      operationId: chequesList
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: List cheques }
  /api/admin/cheques/received:
    post:
      operationId: chequesReceived
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Record received cheque }
  /api/admin/cheques/issued:
    post:
      operationId: chequesIssued
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Record issued cheque }
  /api/admin/cheques/{id}/cashed:
    put:
      operationId: chequesMarkCashed
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Mark cheque cashed }
  /api/admin/cheques/{id}:
    delete:
      operationId: chequesDelete
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Delete cheque }
  /api/admin/whatsapp/status:
    get:
      operationId: waStatus
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: WhatsApp connection status }
  /api/admin/whatsapp/send:
    post:
      operationId: waSend
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Send WhatsApp message }
  /api/admin/whatsapp/send-void-alert:
    post:
      operationId: waSendVoidAlert
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Send bill-void WhatsApp alert }
  /api/admin/whatsapp/reconnect:
    post:
      operationId: waReconnect
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Force WhatsApp reconnect }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT